--- a/frontend/compiled/hint.go
+++ b/frontend/compiled/hint.go
@@ -6,31 +6,42 @@ import (
 	"reflect"
 
 	"github.com/consensys/gnark/backend/hint"
+	"github.com/consensys/gnark/frontend/field"
 	"github.com/fxamacker/cbor/v2"
 )
 
 // Hint represents a solver hint
 // it enables the solver to compute a Wire with a function provided at solving time
 // using pre-defined inputs
-type Hint struct {
+type Hint[E field.El, ptE field.PtEl[E]] struct {
 	ID     hint.ID       // hint function id
 	Inputs []interface{} // terms to inject in the hint function
 	Wires  []int         // IDs of wires the hint outputs map to
 }
 
-func (h Hint) inputsCBORTags() (cbor.TagSet, error) {
+// hintMarshal is like [Hint], but doesn't implement cbor.Marshal, so would use
+// default cbor marshalling. As the inputs in [Hint] may have different types,
+// then have to figure out at runtime how to marshal different types, but for
+// the rest can use standard cbor serialisation.
+type hintMarshal struct {
+	ID     hint.ID
+	Inputs []interface{}
+	Wires  []int
+}
+
+func (h Hint[E, ptE]) inputsCBORTags() (cbor.TagSet, error) {
 	defTagOpts := cbor.TagOptions{EncTag: cbor.EncTagRequired, DecTag: cbor.DecTagRequired}
 	tags := cbor.NewTagSet()
-	if err := tags.Add(defTagOpts, reflect.TypeOf(LinearExpression{}), 25443); err != nil {
+	if err := tags.Add(defTagOpts, reflect.TypeOf(LinearExpression[E, ptE]{}), 25443); err != nil {
 		return nil, fmt.Errorf("new LE tag: %w", err)
 	}
-	if err := tags.Add(defTagOpts, reflect.TypeOf(Term(0)), 25445); err != nil {
+	if err := tags.Add(defTagOpts, reflect.TypeOf(Term[E, ptE]{}), 25445); err != nil {
 		return nil, fmt.Errorf("new term tag: %w", err)
 	}
 	return tags, nil
 }
 
-func (h Hint) MarshalCBOR() ([]byte, error) {
+func (h Hint[E, ptE]) MarshalCBOR() ([]byte, error) {
 	tags, err := h.inputsCBORTags()
 	if err != nil {
 		return nil, fmt.Errorf("cbor tags: %w", err)
@@ -39,8 +50,6 @@ func (h Hint) MarshalCBOR() ([]byte, error) {
 	if err != nil {
 		return nil, err
 	}
-	// v of type vt is Hint but does not implement cbor.Marshaler
-	type vt Hint
 	inputs := make([]interface{}, len(h.Inputs))
 	// map big.Int to bytes
 	for i := range h.Inputs {
@@ -65,11 +74,20 @@ func (h Hint) MarshalCBOR() ([]byte, error) {
 			inputs[i] = h.Inputs[i]
 		}
 	}
-	v := vt{ID: h.ID, Inputs: inputs, Wires: h.Wires}
+	v := hintMarshal{ID: h.ID, Inputs: inputs, Wires: h.Wires}
 	return enc.Marshal(v)
 }
 
-func (h *Hint) UnmarshalCBOR(b []byte) error {
+// v of type vt is Hint but does not implement cbor.Unmarshal. We use this type
+// to unmarshal everything except Inputs field which we have to handle
+// separately as we support different types.
+type hintUnmarshal struct {
+	ID     hint.ID
+	Inputs []cbor.RawTag
+	Wires  []int
+}
+
+func (h *Hint[E, ptE]) UnmarshalCBOR(b []byte) error {
 	tags, err := h.inputsCBORTags()
 	if err != nil {
 		return fmt.Errorf("cbor tags: %w", err)
@@ -78,13 +96,8 @@ func (h *Hint) UnmarshalCBOR(b []byte) error {
 	if err != nil {
 		return fmt.Errorf("decoder: %w", err)
 	}
-	// v of type vt is Hint but does not implement cbor.Marshaler
-	type vt struct {
-		ID     hint.ID
-		Inputs []cbor.RawTag
-		Wires  []int
-	}
-	var v vt
+
+	var v hintUnmarshal
 	if err := dec.Unmarshal(b, &v); err != nil {
 		return fmt.Errorf("unmarshal: %w", err)
 	}
@@ -92,17 +105,17 @@ func (h *Hint) UnmarshalCBOR(b []byte) error {
 	for i, vin := range v.Inputs {
 		switch vin.Number {
 		case 25443:
-			var v []Term
+			var v []Term[E, ptE]
 			if err := dec.Unmarshal(vin.Content, &v); err != nil {
 				return fmt.Errorf("unmarshal linear expression: %w", err)
 			}
-			inputs[i] = LinearExpression(v)
+			inputs[i] = LinearExpression[E, ptE](v)
 		case 25445:
 			var v uint64
 			if err := dec.Unmarshal(vin.Content, &v); err != nil {
 				return fmt.Errorf("unmarshal term: %w", err)
 			}
-			inputs[i] = Term(v)
+			inputs[i] = Term[E, ptE]{Var: v}
 		case 25446:
 			v := new(big.Int)
 			var bb []byte