--- a/frontend/compiled/r1cs_sparse.go
+++ b/frontend/compiled/r1cs_sparse.go
@@ -15,44 +15,45 @@
 package compiled
 
 import (
-	"math/big"
 	"strings"
+
+	"github.com/consensys/gnark/frontend/field"
 )
 
 // R1CS decsribes a set of SparseR1C constraint
-type SparseR1CS struct {
-	ConstraintSystem
-	Constraints []SparseR1C
+type SparseR1CS[E field.El, ptE field.PtEl[E]] struct {
+	ConstraintSystem[E, ptE]
+	Constraints []SparseR1C[E, ptE]
 }
 
 // GetNbConstraints returns the number of constraints
-func (cs *SparseR1CS) GetNbConstraints() int {
+func (cs *SparseR1CS[E, ptE]) GetNbConstraints() int {
 	return len(cs.Constraints)
 }
 
 // SparseR1C used to compute the wires
 // L+R+M[0]M[1]+O+k=0
 // if a Term is zero, it means the field doesn't exist (ex M=[0,0] means there is no multiplicative term)
-type SparseR1C struct {
-	L, R, O Term
-	M       [2]Term
-	K       int // stores only the ID of the constant term that is used
+type SparseR1C[E field.El, ptE field.PtEl[E]] struct {
+	L, R, O Term[E, ptE]    // left, right, output terms. Left and right are addition terms
+	M       [2]Term[E, ptE] // multiplication terms
+	K       E               // constant term
 }
 
-func (r1c *SparseR1C) String(coeffs []big.Int) string {
+func (r1c *SparseR1C[E, ptE]) String() string {
 	var sbb strings.Builder
 	sbb.WriteString("L[")
-	r1c.L.string(&sbb, coeffs)
+	r1c.L.string(&sbb)
 	sbb.WriteString("] * R[")
-	r1c.R.string(&sbb, coeffs)
+	r1c.R.string(&sbb)
 	sbb.WriteString("] + M0[")
-	r1c.M[0].string(&sbb, coeffs)
+	r1c.M[0].string(&sbb)
 	sbb.WriteString("] + M1[")
-	r1c.M[1].string(&sbb, coeffs)
+	r1c.M[1].string(&sbb)
 	sbb.WriteString("] + O[")
-	r1c.O.string(&sbb, coeffs)
+	r1c.O.string(&sbb)
 	sbb.WriteString("] + K[")
-	sbb.WriteString(coeffs[r1c.K].String())
+	sbb.WriteString(ptE(&r1c.K).String())
 	sbb.WriteString("]")
 
 	return sbb.String()