… frontend/cs/r1cs/api_assertions.go
… frontend/cs/r1cs/r1cs_test.go
… frontend/cs/scs/api.go
… frontend/cs/scs/api_assertions.go
… frontend/cs/scs/builder.go
… frontend/field/element.go
… internal/backend/bls12-377/cs/r1cs.go
… internal/backend/bls12-377/cs/r1cs_sparse.go
… internal/backend/bls12-377/cs/solution.go
… internal/backend/bls12-377/groth16/setup.go
… internal/backend/bls12-377/plonk/setup.go
… internal/backend/bls12-377/plonkfri/setup.go
… internal/backend/bls12-381/cs/r1cs.go
… internal/backend/bls12-381/cs/r1cs_sparse.go
… internal/backend/bls12-381/cs/solution.go
… internal/backend/bls12-381/groth16/setup.go
… internal/backend/bls12-381/plonk/setup.go
… internal/backend/bls12-381/plonkfri/setup.go
… internal/backend/bls24-315/cs/r1cs.go
… internal/backend/bls24-315/cs/r1cs_sparse.go
… internal/backend/bls24-315/cs/solution.go
… internal/backend/bls24-315/groth16/setup.go
… internal/backend/bls24-315/plonk/setup.go
… internal/backend/bls24-315/plonkfri/setup.go
… internal/backend/bls24-317/cs/r1cs.go
… internal/backend/bls24-317/cs/r1cs_sparse.go
… internal/backend/bls24-317/cs/solution.go
… internal/backend/bls24-317/groth16/setup.go
… internal/backend/bls24-317/plonk/setup.go
… internal/backend/bls24-317/plonkfri/setup.go