A quick try at parametrising the frontend types.

I ran into problems when working with large linear expressions which appear in zk-unfriendly circuits (for example #401). The problem was that linear expressions are huge and in additions we have to add coefficients wire-by-wire. And this created two issues:

1. the coefficients were stored in a lookup table for avoiding redundancy and having compact terms (in linear expressions). However, when reducing LEs, then some coefficients became unused. But we cannot remove them from the lookup table as the compiler has no knowledge if the inputs are used in other operations. In #401 context it meant that the size of lookup table grew to 10M entries very quickly (after 15K constraints) and lookups became very expensive.

2. we perform a lot of reductions which incur super-linear cost in the size of linear expressions.

So, I was trying to solve both of the issues. To get rid of the lookup tables, I redefined `Term` to include the coefficient directly (not through ID and table). Additionally, I parametrised everything so that the coefficients would be the actual underlying type (`fr.Element` for scalar fields).

For the second problem, I implemented a feature to record a linear expression in the circuit to replace it with a single term. I did this by creating a new variable `h` and then defining constraint `(\sum_{i=0}^{huge} c_i a_i) * 1 = h` and then returning `h` instead. The parameter when to do this is configurable (and the default behaviour is not to "compress" LEs) and can be provided to compiler as `frontend.WithCompressThreshold` option. This provides satisfactory result - for keccak-f permutation with threshold 500 the incurred overhead (in the number of constraints) is around 9%. But compared to the previous situation where I was unable to compile the circuit at all (and now compiles in 30s), is satisfactory (at least for now).

PR is not yet ready:
- [ ] some tests fail (but they seem easy to fix),
- [ ] I leak the type parameters slightly into backend (I'm not sure if it actually hurts, because the type parameters are fixed for the curves. In any case would be easy to fix with code generation)
- [ ] I lost some optimisations in the backend (I use the same `Term` definition as for frontend, but for the backend we can actually build the coefficient table and use it. I also lost batch-invert in plonk backend because lacking coefficient table)
- [ ] lacking documentation for new definitions
- [ ] could use more pool for element reuse (but need to look at the assembly to figure out when allocated on heap and when not)
- [ ] frontend has still a lot of *big.Ints, but want to get rid of them (for constants etc.).

I haven't benchmarked on larger circuits yet, but tests seem to be working not-too-slow (subjectively, at least). @gbotrel, can give some feedback if is worth fixing.