@gbotrel 2022-12-01
Did add couple of comments -- not sure this patch is not hurting other cases; for example, in `std/groth16_bls12377`, running `BenchmarkCompile` against develop:

```
benchmark               old ns/op     new ns/op      delta
BenchmarkCompile-10     394783403     1292244583     +227.33%

benchmark               old allocs     new allocs     delta
BenchmarkCompile-10     2952001        15556641       +426.99%

benchmark               old bytes     new bytes      delta
BenchmarkCompile-10     210047573     1595247472     +659.47%
```