@ivokub 2022-12-01
> Will visit this PR and experiment locally today :)
> 
> > I lost some optimisations in the backend (I use the same Term definition as for frontend, but for the backend we can actually build the coefficient table and use it. I also lost batch-invert in plonk backend because lacking coefficient table)
> 
> --> that's a must have 👍

Definitely! I think actually that this would be a great opportunity to fully split frontend and backend. Make types `frontend/compiled` private for frontend only and then define the corresponding non-generic generated types in backend (with the table-based coefficients etc). Then can easily reintroduce the optimisations I removed.