@ivokub 2022-12-14
This PR is superseded by #412. Keeping it now for porting bda92771794405ebbd43dbb482d6c6443bc8d7b9 on top.